pyproject.toml
[project]
name = "friday-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "openai>=2.8.0",
    "pydantic-ai>=1.18.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "textual>=6.6.0",
    "textual-dev>=1.8.0",
    "typer>=0.20.0",
]
